post

+++ common/models/post.go
@@ -6,8 +6,8 @@ import (
 
 type Post struct {
 	gorm.Model
-	SectionID uint   `gorm:"column:section_id;not null;index" json:"sectionId"`
-	AuthorID  uint   `gorm:"column:author_id;not null;index" json:"authorId"`
+	SectionID int    `gorm:"column:section_id;not null;index" json:"sectionId"`
+	AuthorID  int64  `gorm:"column:author_id;not null;index" json:"authorId"`
 	Title     string `gorm:"column:title;size:255;not null" json:"title"`
 	Content   string `gorm:"column:content;type:text" json:"content"`
 	Pinned    bool   `gorm:"column:pinned;not null;default:false" json:"pinned"`
